+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Next
+
+- **[Feature]** Detect more environments: Deepin DE, EDE, Endless OS, Hyprland, Old (legacy menus), Pantheon (Elementary OS), Razor, ROX, Sway, Trinity DE ([@nagua](https://github.com/nagua), [#6](https://github.com/demurgos/detect-desktop-environment/pull/6))
+- **[Fix]** Fix parsing of `XDG_CURERNT_DESKTOP` ([@nagua](https://github.com/nagua), [#6](https://github.com/demurgos/detect-desktop-environment/pull/6))
+
 # 1.0.0 (2023-03-18)
 
 - **[Breaking change]** Mark the enum `DesktopEnvironment` as non-exhaustive. This allows adding new environments in